fixes & add BoardGroup component
