+++ src/pages/HomePage.tsx
@@ -1,33 +1,14 @@
 import React from "react";
-import { Container, Stack, Typography } from "@mui/material";
-import useAppSelector from "../hooks/useAppSelector";
-import { IBoard } from "../interfaces/IBoard";
-import useDebounce from "../hooks/useDebounce";
-import SearchFilter from "../components/UI/SearchFilter/SearchFilter";
-import { BoardForm, Board } from "../components";
+import { Container } from "@mui/material";
+import { BoardForm } from "../components";
+import BoardsGroup from "../components/BoardsGroup/BoardsGroup";
 
 const HomePage: React.FC = () => {
-    const boards = useAppSelector((state) => state.boardReducer.boards);
-    const [filteredBoards, setFilteredBoards] = React.useState<IBoard[]>(boards);
-
-    const [filterValue, setFilterValue] = React.useState<string>("");
-    const debouncedValue = useDebounce(filterValue, 500);
-
     const [isExpanded, setIsExpanded] = React.useState<boolean>(false);
 
     const onChangeFormExpanded = React.useCallback(() => {
         setIsExpanded(!isExpanded);
     }, [isExpanded]);
-
-    const onChangeFilterValue = React.useCallback((value: string) => {
-        setFilterValue(value);
-    }, []);
-
-    React.useEffect(() => {
-        setFilteredBoards(
-            boards.filter((board) => board.title.toLocaleLowerCase().includes(debouncedValue))
-        );
-    }, [debouncedValue, boards]);
 
     return (
         <Container
@@ -53,29 +34,7 @@ const HomePage: React.FC = () => {
                     boxShadow: 4,
                 }}
             >
-                <SearchFilter inputValue={filterValue} onChangeHandler={onChangeFilterValue} />
-                <Stack
-                    direction="column"
-                    justifyContent="flex-start"
-                    alignItems="stretch"
-                    spacing={2}
-                    sx={{ padding: "15px 0" }}
-                >
-                    {filteredBoards.length ? (
-                        filteredBoards.map((board) => <Board key={board.id} board={board} />)
-                    ) : (
-                        <Typography
-                            variant={"h5"}
-                            sx={{
-                                textAlign: "center",
-                                color: "#fff",
-                                fontWeight: "bold",
-                            }}
-                        >
-                            Ничего не найдено
-                        </Typography>
-                    )}
-                </Stack>
+                <BoardsGroup />
             </Container>
         </Container>
     );
